% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gene_element.R
\name{draw_gene_arrow}
\alias{draw_gene_arrow}
\title{Draws a gene arrow to the current viewport, in the native coordinate system.
Not for export.}
\usage{
draw_gene_arrow(start, end, fill)
}
\description{
Draws a gene arrow to the current viewport, in the native coordinate system.
Not for export.
}

