% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_element.R
\name{gene_element}
\alias{gene_element}
\title{Genes drawn as arrows}
\usage{
gene_element(data = NULL, mapping = NULL)
}
\arguments{
\item{data}{Data frame of genes to be drawn. If NULL, will inherit from the
default set with ‘genemap’}

\item{mapping}{Named character vector giving the mapping between data columns
and element aesthetics. If NULL, will inherit from the default set with
‘genemap’. Mappings set for this element will overwrite any existing default
mappings.}
}
\description{
This plot element will draw genes as arrows on a string.
}
\seealso{
Other genemapr: \code{\link{genemap}}
}

