% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genemap.R
\name{genemap}
\alias{genemap}
\title{Draw a gene map}
\usage{
genemap(data = NULL, mapping = NULL)
}
\arguments{
\item{data}{Data frame containing default data to be used in drawing the gene
map. This default will be used by any plot element that is not given an
explicit dataset. Can be left as NULL.}

\item{mapping}{Default mapping between columns in data and plot aesthetics.
Must be a named character vector. Can be left as NULL.}
}
\description{
The ‘genemapr’ package uses a ggplot2-like paradigm to draw maps of genes on
a molecule.
}
\seealso{
Other genemapr: \code{\link{gene_element}}
}

