% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.genemap.R
\name{+.genemap}
\alias{+.genemap}
\title{Overload `+` operator to allow genemapr elements to be easily combined. Not
for export.}
\usage{
\method{+}{genemap}(genemap_obj, element)
}
\description{
Overload `+` operator to allow genemapr elements to be easily combined. Not
for export.
}

