% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.gene_element.R
\name{print.gene_element}
\alias{print.gene_element}
\title{Print function for gene elements. Not for export.}
\usage{
\method{print}{gene_element}(element)
}
\description{
Print function for gene elements. Not for export.
}

