% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.genemap.R
\name{print.genemap}
\alias{print.genemap}
\title{Print function for the plot object. Sets up plot area. Not for export.}
\usage{
\method{print}{genemap}(genemap_obj)
}
\description{
Print function for the plot object. Sets up plot area. Not for export.
}

